Move ReadWriteHandle-related stuff into handles.py.
