+++ apis/python/src/tiledbsoma/util_tiledb.py
@@ -1,16 +1,12 @@
-from typing import Any, Callable, Generic, MutableMapping, Optional, TypeVar, cast
+from typing import TypeVar, cast
 
-import attrs
 import numpy as np
 import pandas._typing as pdt
 import scipy.sparse as sp
 import tiledb
 from pandas.api.types import infer_dtype, is_categorical_dtype
-from somacore import options
 
-from .exception import DoesNotExistError, SOMAError
-from .options import SOMATileDBContext
-from .types import NPNDArray, PDSeries, TDBHandle
+from .types import NPNDArray, PDSeries
 
 SOMA_OBJECT_TYPE_METADATA_KEY = "soma_object_type"
 SOMA_ENCODING_VERSION_METADATA_KEY = "soma_encoding_version"
@@ -105,115 +101,3 @@ def is_duplicate_group_key_error(e: tiledb.TileDBError) -> bool:
         return True
 
     return False
-
-
-_TDBO = TypeVar("_TDBO", bound=TDBHandle)
-_TDBO_co = TypeVar("_TDBO_co", bound=TDBHandle, covariant=True)
-_Self = TypeVar("_Self")
-
-
-# TODO: We should only need to keep around one of the read or write handles
-# after we do the initial open (where we read metadata and group contents).
-# For now, we're leaving this as-is.
-@attrs.define()
-class ReadWriteHandle(Generic[_TDBO_co]):
-    """Paired read/write handles to a TileDB object.
-
-    You can (and should) access the members, but do not reassign them.
-    """
-
-    uri: str
-    context: SOMATileDBContext
-    reader: _TDBO_co
-    maybe_writer: Optional[_TDBO_co]
-    closed: bool = False
-
-    @classmethod
-    def open_array(
-        cls,
-        uri: str,
-        mode: options.OpenMode,
-        context: SOMATileDBContext,
-    ) -> "ReadWriteHandle[tiledb.Array]":
-        return cls._open(tiledb.open, uri, mode, context)
-
-    @classmethod
-    def open_group(
-        cls,
-        uri: str,
-        mode: options.OpenMode,
-        context: SOMATileDBContext,
-    ) -> "ReadWriteHandle[tiledb.Group]":
-        return cls._open(tiledb.Group, uri, mode, context)
-
-    @classmethod
-    def _open(
-        cls,
-        tiledb_cls: Callable[..., _TDBO],
-        uri: str,
-        mode: options.OpenMode,
-        context: SOMATileDBContext,
-    ) -> "ReadWriteHandle[_TDBO]":
-        try:
-            rd = tiledb_cls(uri, "r", ctx=context.tiledb_ctx)
-            try:
-                wr = (
-                    tiledb_cls(uri, "w", ctx=context.tiledb_ctx)
-                    if mode == "w"
-                    else None
-                )
-            except Exception:
-                rd.close()
-                raise
-        except tiledb.TileDBError as tdbe:
-            if is_does_not_exist_error(tdbe):
-                raise DoesNotExistError(
-                    f"{tiledb_cls.__name__} {uri!r} does not exist"
-                ) from tdbe
-            raise
-        return ReadWriteHandle(uri, context, reader=rd, maybe_writer=wr)
-
-    @property
-    def writer(self) -> _TDBO_co:
-        if self.maybe_writer is None:
-            raise SOMAError(f"cannot write to {self.uri!r} opened in read-only mode")
-        return self.maybe_writer
-
-    @property
-    def mode(self) -> options.OpenMode:
-        return "r" if self.maybe_writer is None else "w"
-
-    @property
-    def metadata(self) -> MutableMapping[str, Any]:
-        if self.maybe_writer is not None:
-            return self.maybe_writer.meta  # type: ignore[no-any-return]
-        return self.reader.meta  # type: ignore[no-any-return]
-
-    def close(self) -> None:
-        if self.closed:
-            return
-        self.closed = True
-        self.reader.close()
-        if self.maybe_writer is not None:
-            self.maybe_writer.close()
-
-    def _flush_hack(self) -> None:
-        """Flushes writes. Use rarely, if ever."""
-        if self.maybe_writer is None:
-            return
-        self.maybe_writer.close()
-        if isinstance(self.maybe_writer, tiledb.Array):
-            self.maybe_writer = tiledb.open(self.uri, "w", ctx=self.context.tiledb_ctx)
-            self.reader.reopen()
-        else:
-            self.maybe_writer = tiledb.Group(self.uri, "w", ctx=self.context.tiledb_ctx)
-            self.reader = tiledb.Group(self.uri, "r", ctx=self.context.tiledb_ctx)
-
-    def __enter__(self: _Self) -> _Self:
-        return self
-
-    def __exit__(self, *_: Any) -> None:
-        self.close()
-
-    def __del__(self) -> None:
-        self.close()
